requirements.txt
# LunarCrush API Integration - Python Dependencies
# Core requirements for social analytics and grant evaluation

# API and HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Optional: Enhanced data visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Optional: Jupyter notebook support
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Type checking and development
mypy>=1.5.0
black>=23.0.0
flake8>=6.0.0

# Optional: Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional: Async support for high-volume requests
aiohttp>=3.8.0
asyncio-throttle>=1.0.0
